requirements.txt
asgiref==3.2.5
Django==2.2.11
django-classy-tags==1.0.0
django-cms==3.7.1
django-filter==2.2.0
django-formtools==2.2
django-jsonfield==1.4.0
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangorestframework==3.11.0
Markdown==3.2.1
pkg-resources==0.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1